[CL2-6567] date block behavior
