+++ front/app/components/EventCard/DateBlocks/DateBlockSingleYear.tsx
@@ -1,0 +1,52 @@
+import React, { memo } from 'react';
+
+// components
+import {
+  EventDateBlock,
+  EventDate,
+  EventMonth,
+  EventDay,
+  EventYear,
+  Separator,
+} from './styling';
+
+interface Props {
+  startAtDay: string;
+  endAtDay?: string;
+  startAtMonth: string;
+  endAtMonth?: string;
+  startAtYear: string;
+  isMultiDayEvent?: boolean;
+}
+
+export default memo<Props>((props) => {
+  const {
+    startAtDay,
+    endAtDay,
+    startAtMonth,
+    endAtMonth,
+    startAtYear,
+    isMultiDayEvent,
+  } = props;
+
+  const isMultiMonthEvent = startAtMonth !== endAtMonth;
+
+  return (
+    <EventDateBlock>
+      <EventDate>
+        <EventMonth>{startAtMonth}</EventMonth>
+        <EventDay>{startAtDay}</EventDay>
+        {isMultiDayEvent && (
+          <>
+            <Separator>-</Separator>
+            {isMultiMonthEvent && <EventMonth>{endAtMonth}</EventMonth>}
+            <EventDay>{endAtDay}</EventDay>
+          </>
+        )}
+      </EventDate>
+      <EventYear>
+        <span>{startAtYear}</span>
+      </EventYear>
+    </EventDateBlock>
+  );
+});
